Updated structures for Jenga version
Snapping, grid-based solution was too daunting. Straw the broke my back was attempting to define whether a specific-shaped relic could be placed on a surface by querying local Spots.
I nearly had a couple solutions, but it was so outside my knowledge-base that I decided I couldn't make it work in the time limit.

Jenga!!